ACHO QUE É O PENÚLTIMO COMMIT
Criado a uri para fechar uma demanda
criado uri para abrir uma demanda
criado o uri para que seja deletado o atendimento de um produtor, caso ele tenha atendido errado
alterado o formato da coluna quantidade, para poder informar casas decimais

+++ src/main/java/br/com/api_eco_feira/model/central/Demanda_Produto_Associados.java
@@ -35,7 +35,7 @@ public class Demanda_Produto_Associados {
     private double valorPrefeitura;
 
     @Column(nullable = false)
-    private int quantidade;
+    private double quantidade;
 
     @Column(nullable = false)
     private double saldo;
